2507: Migrate RegionController to SpringBoot
