+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/graphql/auth/ViewAdministratorsQueryController.kt
@@ -12,7 +12,6 @@ import app.ehrenamtskarte.backend.shared.exceptions.ForbiddenException
 import app.ehrenamtskarte.backend.shared.exceptions.UnauthorizedException
 import com.expediagroup.graphql.generator.annotations.GraphQLDescription
 import graphql.schema.DataFetchingEnvironment
-import org.dataloader.DataLoaderRegistry
 import org.jetbrains.exposed.sql.SortOrder
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import org.jetbrains.exposed.sql.and
@@ -20,29 +19,13 @@ import org.jetbrains.exposed.sql.not
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.springframework.graphql.data.method.annotation.Argument
 import org.springframework.graphql.data.method.annotation.QueryMapping
-import org.springframework.graphql.execution.BatchLoaderRegistry
-import org.springframework.graphql.execution.DefaultBatchLoaderRegistry
 import org.springframework.stereotype.Controller
-import app.ehrenamtskarte.backend.graphql.regions.types.Region
-import org.springframework.graphql.data.method.annotation.SchemaMapping
-import java.awt.print.Book
-import java.util.concurrent.CompletableFuture
-
 
 @Controller
-class ViewAdministratorsQueryController(
-    private val registry: BatchLoaderRegistry,
-) {
+class ViewAdministratorsQueryController {
     @GraphQLDescription("Returns the requesting administrator as retrieved from his JWT token.")
     @QueryMapping
     fun whoAmI(dfe: DataFetchingEnvironment): Administrator {
-        val regionLoader = dfe.getDataLoader<Int, Region>(Region::class.java.getName())
-
-        // val dataLoaderRegistry = DataLoaderRegistry.newRegistry().build()
-       // registry.registerDataLoaders(dataLoaderRegistry, dfe.graphQlContext)
-       // val loader = dataLoaderRegistry.getDataLoader<Int, Region>(Region::class.java.simpleName)
-       // val dataLoader = dfe.getDataLoader<Int, Region>(Region::class.java.simpleName)
-
         val authContext = dfe.requireAuthContext()
         val admin = authContext.admin
 
@@ -52,12 +35,6 @@ class ViewAdministratorsQueryController(
             }
             Administrator.fromDbEntity(admin)
         }
-    }
-
-    @SchemaMapping
-    fun region(admin: Administrator, dfe: DataFetchingEnvironment): CompletableFuture<Region> {
-        val regionLoader = dfe.getDataLoader<Int, Region>(Region::class.java.getName())
-        return regionLoader!!.load(admin.regionId)
     }
 
     @GraphQLDescription(
